+++ thinkjoy-mu-common/src/main/java/cn/thinkjoy/common/managerui/iauth/client/handler/AccessTokenHandler.java
@@ -134,7 +134,7 @@ public class AccessTokenHandler extends AbstractTokenHandler {
         }
 
         AccessToken accessToken = new AccessToken();
-        accessToken.setUserId(user.getId());
+        accessToken.setUserId((Long)user.getId());
 
         tokenStore.store(accessToken.getValue(), accessToken);
